simplify build (#11)
Remove build.rs for just a standard relative path include.

build scripts are an advanced feature. It just happened that this
particular crate did not play nice with a non-cargo build system.

Likely it would be possible to fix it locally, but I actually don't
see any reason for this crate to have a build.rs

I checked it locally and it seems to produce exactly the same results.

LMK if I am missing something.